refactor: fix placement of builder helpers, add unit tests, improve docs
1. Move build_decoder_from_embeds and build_embedding_from_features from
   _cache_utils.py to _base.py where graph-building helpers belong.
   _cache_utils.py now contains only KV cache and linear-attention utilities.
   Update the 3 callers to import from _base. Export both helpers from
   mobius.tasks.__init__.

2. Add comment in build_decoder_from_embeds explaining why _cache_utils
   is imported inside the function body (namespace clarity — no circular
   dependency exists after the move).

3. Fix SpeechToTextTask.build() to call _validate_components() before
   the isinstance config check, consistent with all other tasks.

4. Add unit tests for both shared helpers in _task_test.py:
   - TestBuildDecoderFromEmbeds: return type, graph name, I/O names, MRoPE shape
   - TestBuildEmbeddingFromFeatures: return type, graph name, I/O names, custom feature name

5. Update _cache_utils.py module docstring to accurately describe the
   (now narrower) scope: KV cache and linear attention utilities only.
