+++ src/mobius/tasks/__init__.py
@@ -49,13 +49,20 @@ __all__ = [
     "VAETask",
     "VideoDenoisingTask",
     "VisionLanguageTask",
+    "build_decoder_from_embeds",
+    "build_embedding_from_features",
     "get_task",
 ]
 
 from mobius._constants import OPSET_VERSION
 from mobius.tasks._adapter import AdapterTask
 from mobius.tasks._audio_feature_extraction import AudioFeatureExtractionTask
-from mobius.tasks._base import ComponentSpec, ModelTask
+from mobius.tasks._base import (
+    ComponentSpec,
+    ModelTask,
+    build_decoder_from_embeds,
+    build_embedding_from_features,
+)
 from mobius.tasks._causal_lm import (
     CausalLMTask,
     HybridCausalLMTask,
