+++ src/mobius/tasks/_base.py
@@ -181,3 +181,163 @@ class ModelTask(ABC):
             A :class:`ModelPackage` containing the built model(s).
         """
         ...
+
+
+# ---------------------------------------------------------------------------
+# Shared graph-builder helpers for multi-component tasks
+# ---------------------------------------------------------------------------
+
+
+def build_decoder_from_embeds(
+    decoder,
+    config: BaseModelConfig,
+    *,
+    mrope: bool = False,
+    hybrid: bool = False,
+) -> ir.Model:
+    """Build an ``inputs_embeds → logits + KV cache`` decoder ONNX graph.
+
+    This is the shared implementation for the ``_build_decoder`` method that
+    is common to :class:`VisionLanguageTask`, :class:`QwenVLTask`,
+    :class:`HybridQwenVLTask`, :class:`SpeechLanguageTask`, and
+    :class:`Phi4MMMultiModalTask`.
+
+    Args:
+        decoder: The decoder sub-module to invoke.
+        config: Architecture configuration.
+        mrope: If ``True``, uses 3D MRoPE position_ids
+            ``[3, batch, seq_len]`` instead of the standard
+            ``[batch, seq_len]``.
+        hybrid: If ``True``, uses hybrid KV + DeltaNet cache inputs/outputs
+            (for Qwen3.5-VL and similar).  Requires ``config.layer_types``.
+
+    Returns:
+        A built :class:`ir.Model` for the decoder.
+    """
+    # Import here rather than at module top to keep _base.py focused on base
+    # class definitions.  _cache_utils does NOT import from _base.py, so there
+    # is no circular dependency — this is purely a namespace-clarity choice.
+    from mobius.tasks._cache_utils import (
+        _make_hybrid_cache_inputs,
+        _make_kv_cache_inputs,
+        _register_hybrid_cache_outputs,
+        _register_kv_cache_outputs,
+        _register_linear_attention_functions,
+    )
+
+    batch = ir.SymbolicDim("batch")
+    seq_len = ir.SymbolicDim("sequence_len")
+    past_seq_len = ir.SymbolicDim("past_sequence_len")
+
+    inputs_embeds = ir.Value(
+        name="inputs_embeds",
+        shape=ir.Shape([batch, seq_len, config.hidden_size]),
+        type=ir.TensorType(config.dtype),
+    )
+    attention_mask = ir.Value(
+        name="attention_mask",
+        shape=ir.Shape([batch, "past_seq_len + seq_len"]),
+        type=ir.TensorType(ir.DataType.INT64),
+    )
+    # MRoPE: 3D position IDs (temporal, height, width) — shape [3, batch, seq_len]
+    # Standard: shape [batch, seq_len]
+    position_ids = ir.Value(
+        name="position_ids",
+        shape=ir.Shape([3, batch, seq_len] if mrope else [batch, seq_len]),
+        type=ir.TensorType(ir.DataType.INT64),
+    )
+
+    graph_inputs = [inputs_embeds, attention_mask, position_ids]
+
+    if hybrid:
+        cache_inputs, past_key_values = _make_hybrid_cache_inputs(
+            config,
+            config.dtype,
+            batch,
+            past_seq_len,
+        )
+    else:
+        cache_inputs, past_key_values = _make_kv_cache_inputs(
+            config.num_hidden_layers,
+            config.num_key_value_heads,
+            config.head_dim,
+            config.dtype,
+            batch,
+            past_seq_len,
+        )
+    graph_inputs.extend(cache_inputs)
+
+    graph, builder = _make_graph(graph_inputs, name="decoder")
+    logits, present_key_values = decoder(
+        builder.op,
+        inputs_embeds=inputs_embeds,
+        attention_mask=attention_mask,
+        position_ids=position_ids,
+        past_key_values=past_key_values,
+    )
+
+    logits.name = "logits"
+    graph.outputs.append(logits)
+
+    if hybrid:
+        _register_hybrid_cache_outputs(
+            graph,
+            present_key_values,
+            config.layer_types or [],
+        )
+        model = _make_model(graph)
+        _register_linear_attention_functions(model, config)
+        return model
+    else:
+        _register_kv_cache_outputs(graph, present_key_values)
+        return _make_model(graph)
+
+
+def build_embedding_from_features(
+    embedding,
+    config: BaseModelConfig,
+    *,
+    feature_name: str,
+    feature_dim: int,
+) -> ir.Model:
+    """Build an ``input_ids + features → inputs_embeds`` embedding ONNX graph.
+
+    This is the shared implementation for ``_build_embedding`` in
+    :class:`VisionLanguageTask` (image features) and
+    :class:`SpeechLanguageTask` (audio features).
+
+    Args:
+        embedding: The embedding sub-module to invoke.
+        config: Architecture configuration.
+        feature_name: Name of the second input (e.g. ``"image_features"`` or
+            ``"audio_features"``).
+        feature_dim: Feature dimension for the second input's last axis.
+
+    Returns:
+        A built :class:`ir.Model` for the embedding model.
+    """
+    batch = ir.SymbolicDim("batch")
+    seq_len = ir.SymbolicDim("sequence_len")
+    num_feature_tokens = ir.SymbolicDim("num_feature_tokens")
+
+    input_ids = ir.Value(
+        name="input_ids",
+        shape=ir.Shape([batch, seq_len]),
+        type=ir.TensorType(ir.DataType.INT64),
+    )
+    features = ir.Value(
+        name=feature_name,
+        shape=ir.Shape([num_feature_tokens, feature_dim]),
+        type=ir.TensorType(config.dtype),
+    )
+
+    graph, builder = _make_graph([input_ids, features], name="embedding")
+    inputs_embeds = embedding(
+        builder.op,
+        input_ids=input_ids,
+        **{feature_name: features},
+    )
+
+    inputs_embeds.name = "inputs_embeds"
+    graph.outputs.append(inputs_embeds)
+    return _make_model(graph)
